break apart primary MainWindow tabs to separate UserControls
